x11vnc: Fix shm close on early shutdown.
Global structures {scanline,fullscreen,scanrect}_shm are initialized to
zero, including the .shmid field. This creates problems when shutdown /
clean_shm is called before these structures are properly initialized in
initialize_polling_images. shm_delete will be called on the structure,
and since the shmid isn't -1, shmctl(id, IPC_RMID...) will be called for
id=0. id=0 is in fact a valid shmid, and it might belong to some other
shared memory for some other processes, creating many potential problems.

Fix is to initialize shmid to -1 to indicate that it is in fact
uninitialized.

BUG=chrome-os-partner:16097
TEST=build factory image, confirm that "aplay" works without error.

Change-Id: I18f3ac50c52818e7c201fc85e8f4df24554237cc
Reviewed-on: https://gerrit.chromium.org/gerrit/37971
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Tested-by: Shawn Nematbakhsh <shawnn@google.com>