linux-info: Avoid calling getfilevar when possible
It can be a bit slow to actually _run_ the kernel Makefile.  Running
it 5 times just to get the version seems overkill.  Avoid it whenever
possible.

It turns out that the code was already assuming that VERSION could be
obtained without running the makefile.  Since PATHLEVEL, SUBLEVEL, etc
are all similar it seems reasonable to assume we can get them
directly.

It also turns out that we don't actually need to get KBUILD_OUTPUT if
it was passed in, so avoid that too.

BUG=chromium:253588
TEST=Time of a no-op incremental kernel build goes down 1.553 seconds
with just this change (from 28.775 seconds to 27.222 seconds).

Change-Id: I4e7f1d73cbb8b8e9622bf0a68e25e15d40b6a5d8
Reviewed-on: https://chromium-review.googlesource.com/59870
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>