Patch smdk-dltool to avoid ENOMEM on large transfers
The USB A-A Exynos5 firmware uploader has an issue where it fails if the
uploaded U-Boot image is too large (happens most often with a flasher
image). It seems that the USB devio kernel interface is very restrictive
on the maximum size it allows for a single URB, and the ENOMEM error it
throws if that is exceeded is not gracefully handled in libusb.

This is really libusb's fault, which should transparently split
transfers into smaller chunks until they pass in this case. However,
since I'm not very inclined to redesign the whole library in a portage
patch, I'll just hack the already pretty ravaged smdk-dltool code to
spoon-feed its data in 256KB chunks instead.

BUG=chrome-os-partner:17860
TEST=Run 'crosfw -VF' with a peach_pit hooked up to servod and A-A, make
sure it succeeds.

Change-Id: I1679f75b687082ee73e6b7e857eb0d734c4d9853
Reviewed-on: https://gerrit.chromium.org/gerrit/61797
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Julius Werner <jwerner@chromium.org>