force update of various perl dependent packages
This stack of changes includes an upgrade of perl. The stage3 we use to
bootstrap the sdk is oldish, but contains some packages which have not
seen updates since. They include deps on perl via the perl eclasses that
the newer version of perl does not provide. While the eclasses have been
updated, the installed versions don't have their deps updated automatically,
so we have to manually revbump them.

Specifically:
 - perl-module.eclass used to have:
   DEPEND="dev-lang/perl[-build]"
 - >=perl-5.16 versions no longer have IUSE=build
 - thus the newer perl-module.eclass has:
   DEPEND="|| ( >=dev-lang/perl-5.16 <dev-lang/perl-5.16[-build] )"
 - old installs of packages still require [-build] in their vdb as they
   were installed with the old DEPEND/RDEPEND settings
 - portage uses the dependency information of the installed packages only,
   not what the current ebuilds in the tree allow
 - trying to install >=perl-5.16 fails due to a portage error:
   emerge: there are no ebuilds built with USE flags to satisfy "dev-lang/perl[-build]".
   !!! One of the following packages is required to complete your request:
   - dev-lang/perl-5.16.3::portage-stable (Missing IUSE: build)

We also need to force a revbump of these packages for existing installs
on developer systems.

BUG=chromium:381483
TEST=See CL:202789

Change-Id: Ibef0d2986858f4a726f4fc054ed72536f2dd8570
Reviewed-on: https://chromium-review.googlesource.com/203010
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>