sys-fs/lvm2: Don't install lvm binaries when USE=device-mapper-only.
Upstream USE flag "device-mapper-only" doesn't install the lvm2*so
libraries but also doesn't install the userland lvm setup binaries
and dm-event related binaries and libraries. The lvm binaries don't
require the libraries, but are not used since we don't support lvm
devices.

This patch further backports the "device-mapper-only" USE flag from
version 2.02.105-r2 to not install the lvm binaries and only install
the device-mapper. This means that the following list of files will
not be installed when using this flag:
/sbin/lv*
/sbin/pv*
/sbin/vg*
/sbin/dmeventd
/usr/sbin/lvm2create_initrd
/lib/udev/rules.d/11-dm-lvm.rules
/lib64/device-mapper/libdevmapper-event-lvm2*so
/lib64/libdevmapper-event*so

With this patch, the installed list matches upstream behaviour.

BUG=chromium:387317
TEST=cbuildbot {x86,amd64,arm}-generic-full
TEST=diff the output of `equery-${BOARD} f lvm2` with and without this patch.

Change-Id: Id011007e3c15735d327415219e7a706d625905e4
Reviewed-on: https://chromium-review.googlesource.com/204977
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>