vim, vim-core: upgraded packages to almost upstream
Upgraded app-editors/vim-core to version 7.3.762 on amd64
Upgraded app-editors/vim to version 7.3.762 on amd64

This resolves an issue with using Python, which I need for a syntax
highlighting plugin I use. Maybe it's useful for others who might run
across the same problem as well. It does not upgrade all the way to the
current stable (7.4.273), since that suffers from
https://bugs.gentoo.org/473372 on cross-compilation.

BUG=None
TEST=Open vi, type ':python echo "Hi!"'. Notice how it doesn't segfault.

Change-Id: I13d3f767d572beb5b642959c21f24779e31881c8
Reviewed-on: https://chromium-review.googlesource.com/216224
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Julius Werner <jwerner@chromium.org>