libp11: Use heap instead of static stack arrays
the libp11 pkcs11_init_cert() function used stack-allocated fixed
size arrays when puling variable length attributes.  This would
cause "Buffer too small" errors with valid (but large) certificate
data.

This CL patches libp11 to query the data size and allocate a buffer
suitable for holding this data.  It does so by passing a NULL value
for data and 0 for size to the C_GetAttributeValue function, which
should return successfully and return the size of the data.

This is encapsulated in a p11_attr function, so it can be called
for multiple data types in pkcs11_init_cert.  While here, add a
null termination check for the CKA_LABEL parameter since it is
assumed to be null-terminated in callers to this function.

BUG=chromium:418180
TEST=network_WiFi_SimpleConnect.wifi_check1x_WPA, manual authentication
with large certificate

Change-Id: Ibf8b8af053e150946d8b523d2d3c7c228249462c
Reviewed-on: https://chromium-review.googlesource.com/220189
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>