pngcrush: New upstream package, 1.7.81
Upgraded media-gfx/pngcrush to version 1.7.81 on amd64. This is inteded
to be used on the host only.

This patch relaxes the dependency on libpng to use the version we have
in chromiumos-overlay since the new version is not required and was
also relaxed upstream in a later version of pngcrush.

BUG=chromium:465647
TEST=sudo emerge pngcrush
TEST=Ran pngcrush on the initramfs images.

Change-Id: I7b6c6f0e65ccf05e315715e714a450ec34e78965
Reviewed-on: https://chromium-review.googlesource.com/259917
Trybot-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>