net-print/cups: hack in cross compile support
The html builds rely on running a locally-built binary on the host. Just
kill that off, since we don't need the docs.

[Note: this seems a bit hacky. Alternatives: support a configure flag to
disable doc generation; or build some of CUPS support libraries for
both the target and host.]

Kill off genstrings runs too, since we don't need locale strings on the
target.

BUG=chrome-os-partner:47758,chrome-os-partner:49267
TEST=`emerge-${BOARD} cups` now works for cross compilation!

Change-Id: Id5ce0bd1334e1bed75cedb98be45fa36e8f852f3
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/326362
Reviewed-by: Mike Frysinger <vapier@chromium.org>