net-print/cups-filters: fix issues when removing ghostscript/ijs
When we remove ghostscript-based filters, CUPS complains about a missing
'gstoraster' filter, even if it doesn't need it. The fix is to drop the
gstoraster rules from the MIME conversion rules. Patch accepted here:

  https://bugs.linuxfoundation.org/show_bug.cgi?id=1346

Also, ijs [1] is part of the ghostscript package, so it needs to be dropped
when we drop ghostscript (i.e., USE="-postscript"). Patch accepted here:

  https://bugs.linuxfoundation.org/show_bug.cgi?id=1345

[1] IJS derives from HPIJS (HP InkJet Server) but is not HP-specific.
    According to its README, it is "a protocol for transmission of
    raster page images."

BUG=chrome-os-partner:47758, chrome-os-partner:49267
TEST=check that cups-filters builds, deploys, and doesn't complain in syslog

Change-Id: Ic1370e816df2fbf103a983b01e23298d6cf8cb1a
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/330188
Reviewed-by: Mike Frysinger <vapier@chromium.org>