net-print/cups-filters: fixup --enable-avahi support
The use of $PKGCONFIG is all wrong in configure.ac, causing us to use
the host pkg-config for cross-compilation, which caused us to drop avahi
support even when we use --enable-avahi. Patching in my
fixes, which were accepted here:

  https://bugs.linuxfoundation.org/show_bug.cgi?id=1347

Now cups-filters will honor the passed $PKG_CONFIG, and it also will
error out (instead of quietly dropping support) when --enable-avahi is
passed, but we can't find the avahi-client library.

BUG=chrome-os-partner:47758, chrome-os-partner:49267
TEST=`cups-browsed` with DNSSD browsing enabled

Change-Id: If57aea7cca4b543c55fc68ab18bec8fc21dfc631
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/334485
Reviewed-by: Mike Frysinger <vapier@chromium.org>