python-exec: work around our compiler wrapper
The python-exec ebuild leaks EPYTHON during configure which breaks our
compiler as it relies on python.  This has been fixed in newer versions
of python-exec, but migrating to that requires a lot more work.  Let's
deploy a hack here as it allows for a smoother upgrade transition.

BUG=chromium:600061
TEST=`cbuildbot chromiumos-sdk` works

Change-Id: I8953b0933db29a00d3488c1a51c7fce3381467b8
Reviewed-on: https://chromium-review.googlesource.com/336990
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>