net-print/cups-filters: do path lookup in gstoraster
The cups-filters package makes poor use of autoconf, such that it
sometimes picks up a host binary (e.g., /usr/bin/gs) and uses that path
for execution on the cross-compiled target. This worked OK when we were
pulling in Ghostscript to the SDK, but when it isn't, ./configure falls
back to just using 'gs'.

This would be OK, except that gstoraster doesn't do pathname lookup.
Let's do the pathname lookup now, and we can resolve the weird autoconf
inconsistencies later.

BUG=chromium:655424
TEST=build cups-filters when ghostscript (i.e., /usr/bin/gs) wasn't
     present in the host SDK

Change-Id: I20d94b087892590bf516851a64f823653033c08f
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/398299
Tested-by: Sean Kau <skau@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>