net-print/cups-filters: don't let autoconf path-search for buildhost binaries
This is a bit of a hack to prevent ./configure from doing an
AC_PATH_PROG() search for pdftops, pdftocairo, or gs. For those
searches, autoconf is incorrectly looking for a binary on the build
system, not on the targeted system. This used to happen to work OK, but
it's not correct -- and we recently removed these from the build
toolset.

cups-filters already supports providing the exact path, to skip the
AC_PATH_PROG() search. We can use that now to skip some of these
searches. We should really just patch out the path search entirely
though, since we can do that perfectly well at runtime (these are
executed with execvp()/execvpe() anyway).

One hiccup with all this: configure.ac still tries to run the host
binary (which may or may not exist) like this:

	AC_MSG_CHECKING([whether pdftops supports -origpagesizes])
	AS_IF([`$CUPS_PDFTOPS -h 2>&1 | grep -q -- -origpagesizes`], [
	....
	AC_MSG_CHECKING([whether pdftops supports -r])
	AS_IF([`$CUPS_PDFTOPS -h 2>&1 | grep -q -- '-r '`], [
	....

That's wrong, at least for cross-compilation.

BUG=chromium:654890
TEST=`emerge --unmerge poppler` on build system (i.e., clean slate); make sure
     `emerge-${BOARD} cups-filters` still works (both with USE=postscript and
     USE=-postscript)

Change-Id: I947720c675e3e922f12cf58479b4199b00ad535f
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/398261
Reviewed-by: Sean Kau <skau@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>