libseccomp: upgraded package to upstream
Upgraded sys-libs/libseccomp to version 2.3.1 on amd64, arm, x86

In preparation for upgrading to systemd 232+, >=libseccomp-2.3.1 is
needed. The latest on Gentoo is 2.3.2, but chromiumos' mirror of Gentoo
distfiles is stale and 2.3.2 not present, see chromium:703244.

While that is being figured out, I think it doesn't hurt to upgrade to
2.3.1 first.

BUG=b:34841890
CQ-DEPEND=CL:457164
TEST=`cbuildbot chromiumos-sdk` works
TEST=`cbuildbot {x86-alex,daisy,lakitu}-release` works

Change-Id: I2debe877824d29333832d7f82213d46c59908d92
Reviewed-on: https://chromium-review.googlesource.com/459009
Commit-Ready: Daniel Wang <wonderfly@google.com>
Tested-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>