libcxx: remove llvm dependency.
We want to libcxx as a target package. It has dependency llvm, but we
do not need to install llvm because we already have llvm installed as
a host package.
However, llvm is not in
${BOARD}/etc/portage/profile/package.provided. We cannot add llvm to
it because mesa needs a different version of llvm installed as a target
package, and it does not specify the llvm version in the dependency.

BUG=chromium:620402
TEST=emerge-arm-generic libcxx, llvm is not installed.

Change-Id: I33eeb0ce5ed85847c32d6df5bfdcdd71ede495f6
Reviewed-on: https://chromium-review.googlesource.com/457154
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>