dev-util/cargo: add cargo ebuild for building rust packages
Cargo is the standard build system and package manager of the rust
ecosystem. Adding it here is essential to our packages that use rust.

This CL includes an ebuild for cargo and an eclass for all packages that
make use of cargo, including cargo itself. These files are taken from
upstream portage with necessary deviations annotated.

TEST=sudo emerge cargo, cargo should be on $PATH
BUG=None

Change-Id: Iaa276c51a07dfc89370d910e2c359e42eefca85b
Reviewed-on: https://chromium-review.googlesource.com/506888
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>