dev-util/cunit: Import from upstream gentoo
CUnit is a Unit testing framework for C.
It is a dependency for building the tests/amdgpu in libdrm.
In particular, tests/amdgpu/deadlock_tests.c causes a GPU reset after
timeout by sending a wait for a condition that will never trigger.

This CL imports cunit from upstream gentoo with the following caveats:
 * EAPI=5 instead of 6
 * KEYWORDS="*"

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=b:65368363
TEST=emerge-kahlee cunit
 => builds clean

Change-Id: Id969adde2ba856aac273f6f4f9b6fa3b6bba13c6
Reviewed-on: https://chromium-review.googlesource.com/747125
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>