sys-apps/lvm2: udev_get_udevdir to get_udevdir
udev_get_udevdir has been deprecated for a while.  crrev.com/c/941974
will change it to an error, so switch to the new function.

Since this is portage-stable, this should ideally be an upgrade to the
newer version instead.  However, there have been 148 lvm2 releases since
we last updated and the current upstream stable version of lvm2 corrupts
the stateful partition when the system boots.  Patch out the deprecated
function here as a temporary measure to unblock updates depending on
this while I debug what's wrong with the newer lvm2 versions.

BUG=chromium:810779,chromium:818278
TEST=emerge-kahlee lvm2

Change-Id: I8bca08d6553c3f1e35dfb54982741aa622fad3d7
Reviewed-on: https://chromium-review.googlesource.com/949971
Commit-Ready: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>