dosfstools: upgraded package to upstream
Upgraded sys-fs/dosfstools to version 4.1 on amd64, arm.
This seemed to cause an error when trying to create the vfat filesystem
on one of the partitions of the disk image, such that udev thinks
that the partition is itself partitioned. To fix this we had to change
the build script to ginore this error, and submit this as a separate CL.
See the linked bug for more discussion.

BUG=chromium:901479
TEST=See the bug for more info
CQ-DEPEND=CL:1323474

Change-Id: I9cbbbf8242e9b6073f5ca822eac1d40ff0938801
Reviewed-on: https://chromium-review.googlesource.com/1323910
Commit-Ready: Nick Crews <ncrews@chromium.org>
Tested-by: Nick Crews <ncrews@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>