lxml, python-futures, trollius: upgraded packages to upstream
Upgraded dev-python/lxml to version 4.2.5 on amd64, arm
Upgraded dev-python/trollius to version 2.1 on amd64, arm
Upgraded virtual/python-futures to version 1 on amd64, arm

These packages are all pre-requisites for pyshark and were previously
updated together in:
7d1de7b91bc23496edf764daa95cdbb33555cd27

This uprev adds support for python3_6.

The upstream LXML ebuild was updated to EAPI=7 and a BDEPEND was added.
For now (due to crbug.com/924701) EAPI7+BDEPEND are not supported, so
down-grade to EAPI6 and convert BDEPEND->DEPEND.

BUG=chromium:926296
TEST=PreCQ passes
TEST=chromiumos-sdk tryjob buildbucketId=8915374561970847360

Change-Id: I1ce92133f46716b2bb2a74e1ca1f19be552f0a34
Reviewed-on: https://chromium-review.googlesource.com/1562918
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>