dev-python/pytest: Add logilab-common as blocker
dev-python/logilab-common was removed from the SDK in chromium:1148789
but some users still have it installed. The new version of pytest has a
file collision with that package, so add this blocker so that anyone who
still has the package doesn't hit an error when trying to update their
chroot.

BUG=chromium:934414
TEST=`emerge dev-python/pytest`

Change-Id: I0118e78d1d4d607ac445736b29129bc77f67981e
Reviewed-on: https://chromium-review.googlesource.com/1575623
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>