ccache, shadowman: upgraded packages to upstream
Upgraded dev-util/shadowman to version 2 on amd64
Upgraded dev-util/ccache to version 3.7.1 on amd64

The goal here is to get the build speed improvement for building the
kernel.  Specifially:
* https://github.com/ccache/ccache/pull/365

...and in order for that patch to work properly:
* https://github.com/ccache/ccache/pull/370

The shadowman ebuild is simply needed for new ccache ebuilds.

This version of ccache is marked as "unstable" in upstream portage, so
I picked with
  cros_portage_upgrade --unstable-ok --upgrade --host \
    shadowman ccache-3.7.1

...and upstream this uses EAPI=7 so I had to hack with:
  https://crrev.com/c/1589137

I believe both the picking unstable and the hacking of the EAPI are
considered OK under Chrome OS rules for portage stable picks.

BUG=chromium:767073
TEST=build_packages

Change-Id: Iecd690af32c675e3b73a40a431d5d5c2488eb64d
Reviewed-on: https://chromium-review.googlesource.com/1535521
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>