portage-stable: Add OWNERS file
Enforcement is not enabled yet. This CL is a suggestion of ownership
candidates based on git activity and you are welcome to add other owners
as you see fit. See the following link for more information on the syntax
for OWNERS files in ChromiumOS repositories:
https://gerrit.googlesource.com/plugins/find-owners/+/master/src/main/resources/Documentation/syntax.md

BUG=chromium:748644
TEST=None

Change-Id: Ia3b537a9bf01047dced0c556634aad7dde801088
Reviewed-on: https://chromium-review.googlesource.com/1593720
Commit-Ready: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>