net-libs/libssh2: Drop openssl slot dependency to allow openssl uprev
libssh2 is a dependency of cargo, which is installed as a binary
package via the host-post-cross package set. libssh2 has a dependency
on openssl. It builds against either openssl 1.0.2 or 1.1.1, but the
chosen version is currently propagated into the dependency information
for the binary package. This creates a conflict on openssl uprev in
the SDK, where we only have a binary package that is compatible with
1.0.2. To solve this, this CL temporarily drops the slot dependency
from the ebuild, which will result in the binary package dependency
to allow either package version. Note that the openssl uprev will
temporarily provide shared objects for *both* openssl versions to not
break linkage.

BUG=chromium:994667
TEST=sudo emerge -v1 libssh2, check /var/db/pkg/net-libs/libssh2/DEPEND to not force a designated openssl version

Change-Id: I8ee4dba1717bc6b0b42588f35d64e38a6b8181c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/portage-stable/+/1871553
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>