ghc-package.eclass: Fix ghc path lookups
ghc-getghc() and ghc-getghcpkg() both assume that they can find an
installed binary.  If ghc itself is being rebuilt, there is no such
binary, which causes various invalid commands to be run.  If ghc can't
be found, run "false" instead to absorb the other command-line
arguments.  This shouldn't produce any difference in the final result
because the commands were failing anyway.

This is a temporary fix to unbreak the SDK builders while we figure out
how to resolve this in the long term.

BUG=chromium:1026092
TEST=chromiumos-sdk tryjob

Change-Id: I293db50c035a191097e2f3f5a3e04d7cce13cb9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/portage-stable/+/1924972
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>