dev-haskell/*: Tag inclusion in new ghc release
We pushed ghc-8.0.2-r1 a couple of days ago.  Some ebuilds track which
versions of ghc normally bundle them, so we need to update them as well:
  * cabal
  * transformers

Doing this in portage-stable to unbreak the SDK builders, but we'll be
migrating all of these over to chromiumos-overlay with a GHC upgrade
shortly.

BUG=chromium:1026092
TEST=sudo emerge shellcheck

Change-Id: I0a12cc9dc8c8070b2c64e5bced33be3356750f4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/portage-stable/+/1931020
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>