OWNERS: Add ebuild-reviews alias
The ebuild-reviews@chromium.org alias triggers gwsq to automatically
assign a reviewer.  Add it to the OWNERS file so that people can see it
in the list of suggested owners.  If it goes well, we can remove the
other owners to give stronger encouragement to use it.

BUG=chromium:1042392
TEST=None

Change-Id: Ie2afe523051e430b105a88693e8deb0bf3e8101a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/portage-stable/+/2001530
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>