dropbear: drop old package
This is an old version with known security vulnerabilities.  No one
seems to depend on it, so drop it to avoid accidental usage.  If
anyone wants it in the future, they can upgrade to the latest.

Also drop libtommath as nothing else uses it, and it's old with a
bad license we don't want to support.

BUG=chromium:1107613
TEST=CQ passes

Change-Id: Icd2eb7edbae731a2e399076d2b66a88c8968c9e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/portage-stable/+/2309183
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>