zsh: upgraded package to upstream
Upgraded app-shells/zsh to version 5.8 on amd64

Note: this was previously in chromiumos-overlay at version 5.0.1, but
does not need CrOS-specific modification, so moving back to
portage-stable.  The intent is to remove from chromiumos-overlay once
this lands.

This package is only installed by developers who want it in their
chroot for convenience, not by default in the chroot, nor on any
devices.

BUG=None
TEST=sudo emerge app-shells/zsh
     Run zsh in chroot, observe its normal operation

Change-Id: I1be9be1b3cd67b96d9f015f66974f229bcd9a943
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/portage-stable/+/2510777
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>